backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
firebase-admin>=6.4.0
langchain>=0.1.7
langchain-community>=0.0.20
chromadb>=0.4.22
rank_bm25>=0.2.2
pandas>=2.2.0
PyPDF2>=3.0.1
python-docx>=1.1.0
python-dotenv>=1.0.1
pydantic>=2.6.1
